Get list of all products

+++ Models/Product.cs
@@ -10,8 +10,8 @@ namespace BackendChallengeAPI.Models
         public int Id { get; set; }
         public string Supplier { get; set; }
         public string Name { get; set; }
-        public decimal Rate { get; set; }
-        public decimal DailyStandingCharge { get; set; }
+        public double Rate { get; set; }
+        public double DailyStandingCharge { get; set; }
         public int ContractLength { get; set; }
         public int Renewable { get; set; }
         public string Status { get; set; }
